funkydex-assets/requirements.txt
discord.py==2.5.2
aiohttp
pandas